--- universal-app-gateway/src/main/filters/dev.properties
@@ -1,20 +1,0 @@
-mvn.server.port = 8080
-
-mvn.log.file.path.name = /workspace/logs/app-gateway
-
-mvn.mysql.onl.jdbc.url = jdbc:mysql://dev-db1.universal.me:3306/db_universal_onl?autoReconnect=true&zeroDateTimeBehavior=convertToNull
-mvn.mysql.onl.jdbc.username = ac_user_pvd
-mvn.mysql.onl.jdbc.password = xxxxxx
-
-mvn.mysql.spu.jdbc.url = jdbc:mysql://dev-db1.universal.me:3306/db_universal_spu?autoReconnect=true&zeroDateTimeBehavior=convertToNull
-mvn.mysql.spu.jdbc.username = ac_user_pvd
-mvn.mysql.spu.jdbc.password = xxxxxx
-
-mvn.dubbo.application.environment = develop
-mvn.dubbo.registry.address = zookeeper://dev-zookeeper1.universal.me:2181?timeout=50000
-
-mvn.redis.host = dev-redis1.universal.me
-mvn.redis.port = 6379
-mvn.redis.password =
-
-mvn.log.appenders = <AppenderRef ref="CONSOLE" />
