--- universal-app-gateway/src/main/resources/application.properties
@@ -1,46 +1,0 @@
-logging.level.root=INFO
-logging.level.org=INFO
-logging.level.com.universal=DEBUG
-
-server.port = ${mvn.server.port}
-
-spring.freemarker.settings.auto_import = spring.ftl as spring
-spring.freemarker.settings.classic_compatible = true
-spring.freemarker.settings.locale = zh_CN
-
-spring.jackson.default-property-inclusion=non_null
-spring.jackson.serialization.write_null_map_values=false
-
-mybatis.datasource.default-target = onl
-mybatis.configuration.map-underscore-to-camel-case = true
-
-mybatis.datasource.driver-class-name[onl] = com.mysql.jdbc.Driver
-mybatis.datasource.url[onl] = ${mvn.mysql.onl.jdbc.url}
-mybatis.datasource.user-name[onl] = ${mvn.mysql.onl.jdbc.username}
-mybatis.datasource.password[onl] = ${mvn.mysql.onl.jdbc.password}
-
-mybatis.datasource.driver-class-name[erp] = com.mysql.jdbc.Driver
-mybatis.datasource.url[erp] = ${mvn.mysql.spu.jdbc.url}
-mybatis.datasource.user-name[erp] = ${mvn.mysql.spu.jdbc.username}
-mybatis.datasource.password[erp] = ${mvn.mysql.spu.jdbc.password}
-
-spring.redis.host = ${mvn.redis.host}
-spring.redis.port = ${mvn.redis.port}
-spring.redis.password = ${mvn.redis.password}
-spring.redis.pool.max-active = 100
-spring.redis.pool.max-idle = 10
-
-com.redis.cache.expires[TTL30D] = 2592000
-com.redis.data.key-serializer-prefix = UNIVERSAL-APP-GW
-
-#com.web.session-interceptor.enable = true
-com.web.validation-message-interceptor.enable = false
-com.web.authorization-interceptor.implement-class = com.universal.gateway.support.interceptor.AuthorizationInterceptor
-
-#com.web.log-request-filter.enable = true
-com.web.validate-request-filter.enable = false
-#com.web.validate-request-filter.implement-class = com.universal.gw.support.interceptor.AuthorizationInterceptor
-#com.web.validate-request-filter.bundles[J2FBFOAI6AJQN76U] = ^[3-4]\.[0-9](\.\d+)*, ^(?!iOS).+
-#com.web.validate-request-filter.bundles[APYQ9WATKK98V2EC] = ^[3-4]\.[0-9](\.\d+)*, ^(?:(?i)IOS).+
-
-
