--- universal-app-gateway/src/main/resources/dubbo-consumer.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns:context="http://www.springframework.org/schema/context" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:dubbo="http://code.alibabatech.com/schema/dubbo"
-       xmlns="http://www.springframework.org/schema/beans"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans
-	http://www.springframework.org/schema/beans/spring-beans-4.3.xsd
-	http://code.alibabatech.com/schema/dubbo
-	http://code.alibabatech.com/schema/dubbo/dubbo.xsd
-	http://www.springframework.org/schema/context
-	http://www.springframework.org/schema/context/spring-context-4.3.xsd">
-
-    <dubbo:annotation package="com.universal.gateway" />
-
-    <dubbo:application name="universal-app-gateway-consumer" environment="${dubbo.application.environment}"/>
-    <dubbo:registry protocol="zookeeper" address="${dubbo.registry.address}" timeout="${dubbo.registry.timeout}" />
-
-</beans>
