--- universal-app-gateway/src/main/resources/dubbo.properties
@@ -1,3 +1,0 @@
-dubbo.application.environment = ${mvn.dubbo.application.environment}
-dubbo.registry.address = ${mvn.dubbo.registry.address}
-dubbo.registry.timeout = 10000
