+++ universal-app-gw/src/main/filters/prd.properties
@@ -1,0 +1,20 @@
+mvn.server.port = 8080
+
+mvn.log.file.path.name = /workspace/logs/app-gw
+
+mvn.mysql.onl.jdbc.url = jdbc:mysql://rdsjp011r5p90p8l064x.mysql.rds.aliyuncs.com:3306/db_universal_onl?autoReconnect=true&zeroDateTimeBehavior=convertToNull
+mvn.mysql.onl.jdbc.username = <#DB_ONL_USER_PVD_USERNAME#>
+mvn.mysql.onl.jdbc.password = xxxxxx
+
+mvn.mysql.spu.jdbc.url = jdbc:mysql://rdsjp011r5p90p8l064x.mysql.rds.aliyuncs.com:3306/db_universal_spu?autoReconnect=true&zeroDateTimeBehavior=convertToNull
+mvn.mysql.spu.jdbc.username = <#DB_SPU_USER_PVD_USERNAME#>
+mvn.mysql.spu.jdbc.password = xxxxxx
+
+mvn.dubbo.application.environment = product
+mvn.dubbo.registry.address = zookeeper://zookeeper1.universal.com:2181
+
+mvn.redis.host = 8283289732404248.m.cnsza.kvstore.aliyuncs.com
+mvn.redis.port = 6379
+mvn.redis.password = xxxxx
+
+mvn.log.appenders = <AppenderRef ref="LOGHUB" />
