+++ universal-app-gw/src/main/filters/uat.properties
@@ -1,0 +1,20 @@
+mvn.server.port = 8080
+
+mvn.log.file.path.name = /workspace/logs/app-gw
+
+mvn.mysql.onl.jdbc.url = jdbc:mysql://uat-db1.universal.me:3306/db_universal_onl?autoReconnect=true&zeroDateTimeBehavior=convertToNull
+mvn.mysql.onl.jdbc.username = <#DB_ONL_USER_PVD_USERNAME#>
+mvn.mysql.onl.jdbc.password = xxxxxx
+
+mvn.mysql.spu.jdbc.url = jdbc:mysql://uat-db1.universal.me:3306/db_universal_spu?autoReconnect=true&zeroDateTimeBehavior=convertToNull
+mvn.mysql.spu.jdbc.username = <#DB_SPU_USER_PVD_USERNAME#>
+mvn.mysql.spu.jdbc.password = xxxxxx
+
+mvn.dubbo.application.environment = product
+mvn.dubbo.registry.address = zookeeper://uat-zookeeper1.universal.me:2181
+
+mvn.redis.host = uat-redis1.universal.me
+mvn.redis.port = 6379
+mvn.redis.password =
+
+mvn.log.appenders =
