+++ universal-app-gw/src/main/java/com/universal/gw/Application.java
@@ -1,10 +1,10 @@
-package com.universal.provider.user;
+package com.universal.gw;
 
-import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
 
-@EnableDubbo
+@EnableCaching
 @SpringBootApplication
 public class Application {
