+++ universal-app-gw/src/main/java/com/universal/gw/domain/User.java
@@ -1,0 +1,25 @@
+package com.universal.gw.domain;
+
+import java.io.Serializable;
+
+public class User implements Serializable {
+
+    private String userName;
+    private String password;
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getUserName() {
+        return userName;
+    }
+
+    public void setUserName(String userName) {
+        this.userName = userName;
+    }
+}
