+++ universal-app-gw/src/main/resources/application.properties
@@ -1,0 +1,23 @@
+logging.level.root=INFO
+logging.level.org=INFO
+logging.level.com.universal=DEBUG
+
+server.port = 8080
+
+dubbo.application.environment =  develop
+dubbo.application.qos-enable = false
+dubbo.registry.address = multicast://224.5.6.7:8888
+
+spring.data.redis.host = localhost
+com.data.redis.space = UNIVERSAL-APP-GW
+
+com.web.log-request-filter.enable = true
+com.web.log-request-filter.url-patterns = /**
+
+com.web.validate-request-filter.enable = true
+
+com.web.validation-message-interceptor.enable = true
+
+com.web.permission-interceptor.implemented-class = com.universal.gw.support.interceptor.MyPermissionInterceptor
+
+
