+++ universal-app-gw/src/main/resources/templates/spring.ftl
