+++ universal-app-gw/src/main/scripts/shutdown.sh
