+++ universal-app-gw/src/main/scripts/startup.sh
