+++ universal-spring-boot-starter/spring-boot-starter-cache-redis/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,3 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-redis.cache.starter.boot.spring.com.universal.sprint.boot.RedisCacheAutoConfiguration
+com.universal.spring.boot.starter.cache.RedisCacheAutoConfiguration
