+++ universal-spring-boot-starter/spring-boot-starter-cache-redis/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-redis.cache.starter.boot.spring.com.universal.sprint.boot.RedisCacheAutoConfiguration
+com.universal.spring.boot.starter.cache.RedisCacheAutoConfiguration
