+++ universal-spring-boot-starter/spring-boot-starter-cache-redis/src/test/resources/application.properties
@@ -1,2 +1,1 @@
-spring.data.redis.cluster.nodes = xxxx
-spring.data.redis.password = xxxx
+spring.data.redis.host = localhost
