+++ universal-spring-boot-starter/spring-boot-starter-data-redis/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,3 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-redis.data.starter.boot.spring.com.universal.sprint.boot.RedisDataAutoConfiguration
+com.universal.spring.boot.starter.data.RedisDataAutoConfiguration
