+++ universal-spring-boot-starter/spring-boot-starter-data-redis/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-redis.data.starter.boot.spring.com.universal.sprint.boot.RedisDataAutoConfiguration
+com.universal.spring.boot.starter.data.RedisDataAutoConfiguration
