+++ universal-spring-boot-starter/spring-boot-starter-data-redis/src/test/resources/application.properties
@@ -1,4 +1,3 @@
-spring.data.redis.cluster.nodes = xxxxx
-spring.data.redis.password = xxxxx
+spring.data.redis.host = localhost
 
 com.data.redis.space = TEST-SPACE
