+++ universal-user-provider/pom.xml
@@ -4,7 +4,6 @@
 
     <groupId>com.universal.provider</groupId>
     <artifactId>universal-user-provider</artifactId>
-    <packaging>jar</packaging>
 
     <parent>
         <groupId>com.universal</groupId>
@@ -14,9 +13,14 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.mysql</groupId>
-            <artifactId>mysql-connector-j</artifactId>
-            <version>9.3.0</version>
+            <groupId>com.universal.api</groupId>
+            <artifactId>universal-user-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.h2database</groupId>
+            <artifactId>h2</artifactId>
+            <version>2.3.232</version>
         </dependency>
         <dependency>
             <groupId>org.mybatis</groupId>
@@ -37,11 +41,6 @@
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-spring-boot-starter</artifactId>
             <version>3.3.4</version>
-        </dependency>
-        <dependency>
-            <groupId>com.universal.api</groupId>
-            <artifactId>universal-user-api</artifactId>
-            <version>${project.version}</version>
         </dependency>
     </dependencies>
 </project>
