--- universal-user-provider/src/main/filters/tst.properties
@@ -1,19 +1,0 @@
-mvn.log.file.path.name = /workspace/logs/user-provider
-
-mvn.mysql.onl.jdbc.url = jdbc:mysql://tst-db1.universal.me:3306/db_universal_onl?autoReconnect=true&zeroDateTimeBehavior=convertToNull
-mvn.mysql.onl.jdbc.username = <#DB_ONL_USER_PVD_USERNAME#>
-mvn.mysql.onl.jdbc.password = xxxxxx
-
-mvn.mysql.spu.jdbc.url = jdbc:mysql://tst-db1.universal.me:3306/db_universal_spu?autoReconnect=true&zeroDateTimeBehavior=convertToNull
-mvn.mysql.spu.jdbc.username = <#DB_SPU_USER_PVD_USERNAME#>
-mvn.mysql.spu.jdbc.password = xxxxxx
-
-mvn.dubbo.protocol.port = 30602
-mvn.dubbo.application.environment = test
-mvn.dubbo.registry.address = zookeeper://tst-zookeeper1.universal.me:2181
-
-mvn.redis.host = tst-redis1.universal.me
-mvn.redis.port = 6379
-mvn.redis.password =
-
-mvn.log.appenders =
