+++ universal-user-provider/src/main/resources/application.properties
@@ -2,32 +2,25 @@ logging.level.root=INFO
 logging.level.org=INFO
 logging.level.com.universal=DEBUG
 
-dubbo.application.name=universal-user-provider
-dubbo.protocol.port=${mvn.dubbo.protocol.port}
-dubbo.registry.address=${mvn.dubbo.registry.address}
+dubbo.application.name = universal-user-provider
+dubbo.application.environment =  develop
+dubbo.application.qos-enable = false
 
-spring.datasource.dynamic.enabled=true
-spring.datasource.maste.url = jdbc:mysql://xx.xx.xx.xx:3306/dynamic
-spring.datasource.maste.username = root
-spring.datasource.maste.password = 123456
+dubbo.protocol.port = 30602
+dubbo.registry.address = multicast://224.5.6.7:8888
 
+spring.datasource.dynamic.enabled = true
+spring.datasource.dynamic.primary = master
+spring.datasource.dynamic.datasource.master.url = jdbc:h2:mem:~/testdb
+spring.datasource.dynamic.datasource.master.username = root
+spring.datasource.dynamic.datasource.master.password = 123456
+spring.datasource.dynamic.datasource.master.driver-class-name = org.h2.Driver
 
+#spring.h2.console.enabled = true
+#spring.h2.console.path = /h2
+spring.jpa.hibernate.ddl-auto = update
+mybatis.configuration.map-underscore-to-camel-case = true
+mybatis.mapper-locations = classpath*:com/universal/**/mapper/xml/*.xml
 
-
-mybatis.datasource.default-target=onl
-mybatis.configuration.map-underscore-to-camel-case=true
-mybatis.datasource.driver-class-name[onl]=com.mysql.jdbc.Driver
-mybatis.datasource.url[onl]=${mvn.mysql.onl.jdbc.url}
-mybatis.datasource.user-name[onl]=${mvn.mysql.onl.jdbc.username}
-mybatis.datasource.password[onl]=${mvn.mysql.onl.jdbc.password}
-mybatis.datasource.driver-class-name[spu]=com.mysql.jdbc.Driver
-mybatis.datasource.url[spu]=${mvn.mysql.spu.jdbc.url}
-mybatis.datasource.user-name[spu]=${mvn.mysql.spu.jdbc.username}
-mybatis.datasource.password[spu]=${mvn.mysql.spu.jdbc.password}
-spring.redis.host=${mvn.redis.host}
-spring.redis.port=${mvn.redis.port}
-spring.redis.password=${mvn.redis.password}
-spring.redis.pool.max-active=100
-spring.redis.pool.max-idle=10
-com.redis.cache.expires[TTL30D]=2592000
-com.redis.data.key-serializer-prefix=UNIVERSAL-USER-PROVIDER
+spring.data.redis.host = localhost
+com.data.redis.space = TEST-SPACE
