+++ universal-user-provider/src/main/resources/data.sql
@@ -1,0 +1,1 @@
+INSERT INTO T_USER (PARENT_ID, USER_NAME, LINK_NAME, STATUS) VALUES (99, 'admin', 'admin1', 'Y');
