--- universal-user-provider/src/main/resources/dubbo.properties
@@ -1,6 +1,0 @@
-dubbo.protocol.port = ${mvn.dubbo.protocol.port}
-dubbo.application.environment = ${mvn.dubbo.application.environment}
-dubbo.registry.address = ${mvn.dubbo.registry.address}
-dubbo.registry.timeout = 10000
-dubbo.provider.timeout = 5000
-dubbo.shutdown.timeout = 5000
