--- universal-user-provider/src/main/resources/log4j2.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE xml>
-<Configuration>
-
-	<Properties>
-		<Property name="log.file.name">${mvn.log.file.path.name}</Property>
-	</Properties>
-
-	<Appenders>
-		<Console name="CONSOLE" target="SYSTEM_OUT">
-			<PatternLayout charset="UTF-8" pattern="[%d{yyyy-MM-dd HH:mm:ss,SSS} %5p] [%t] [%c{2}] - %m%n" />
-		</Console>
-		<RollingFile name="FILE" fileName="${log.file.name}.log" filePattern="${log.file.name}.%d{yyyy-MM-dd}.log">
-			<PatternLayout charset="UTF-8" pattern="[%d{yyyy-MM-dd HH:mm:ss,SSS} %5p] [%t] [%c{2}] - %m%n" />
-			<Policies>
-				<TimeBasedTriggeringPolicy interval="1" modulate="true" />
-			</Policies>
-		</RollingFile>
-	</Appenders>
-
-	<Loggers>
-		<AsyncLogger name="com.universal" level="DEBUG" />
-		<AsyncRoot level="DEBUG">
-			<AppenderRef ref="FILE" />${mvn.log.appenders}
-		</AsyncRoot>
-	</Loggers>
-</Configuration>
