+++ universal-user-provider/src/main/resources/schema.sql
@@ -1,0 +1,11 @@
+CREATE TABLE IF NOT EXISTS T_USER (
+    ID INT AUTO_INCREMENT PRIMARY KEY,
+    PARENT_ID INT,
+    USER_NAME VARCHAR(255) NOT NULL,
+    PASSWORD VARCHAR(255),
+    LINK_NAME VARCHAR(255),
+    TYPE VARCHAR(255),
+    STATUS VARCHAR(255) NOT NULL,
+    PHONE VARCHAR(255),
+    CRM_CODE VARCHAR(255)
+);
